+++ spec/rubocop/cli_spec.rb
@@ -55,7 +55,7 @@ Usage: rubocop [options] [file1, file2, ...]
                                        [e]macs
                                        [j]son
                                        [f]iles
-                                       [cop]count
+                                       [o]ffences
                                        custom formatter class name
     -o, --out FILE                   Write output to a file instead of STDOUT.
                                      This option applies to the previously